Separated Attack into own script
Separated Attack functions into their own script
