+++ Iso Adventure/Assets/PlayerCombat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af0649b97e729964f9f133621e2351a1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
