+++ Iso Adventure/Assets/Scenes/SampleScene.unity
@@ -346,11 +346,6 @@ MonoBehaviour:
   ground:
     serializedVersion: 2
     m_Bits: 64
-  basicAttackPoint: {fileID: 0}
-  basicAttackRange: 0.5
-  enemyLayers:
-    serializedVersion: 2
-    m_Bits: 0
 --- !u!1 &672830525 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 5430377791670745391, guid: 0616308f626330740800bf213bc4dd38,
@@ -366,9 +361,14 @@ MonoBehaviour:
   m_GameObject: {fileID: 672830525}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: dad47f945930c7a4594d4e34cf884a7d, type: 3}
+  m_Script: {fileID: 11500000, guid: af0649b97e729964f9f133621e2351a1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  basicAttackPoint: {fileID: 2027668770}
+  basicAttackRange: 0.5
+  enemyLayers:
+    serializedVersion: 2
+    m_Bits: 128
 --- !u!1 &705507993
 GameObject:
   m_ObjectHideFlags: 0
@@ -1342,36 +1342,12 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1 &2027668769
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 2027668770}
-  m_Layer: 0
-  m_Name: BasicAttackPoint
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &2027668770
+--- !u!4 &2027668770 stripped
 Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
+  m_CorrespondingSourceObject: {fileID: 2027668770, guid: 0616308f626330740800bf213bc4dd38,
+    type: 3}
+  m_PrefabInstance: {fileID: 5430377793167155872}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2027668769}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0.701}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 1498520468}
-  m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &2069293315
 GameObject:
   m_ObjectHideFlags: 0
@@ -1414,26 +1390,6 @@ PrefabInstance:
         type: 3}
       propertyPath: m_Name
       value: Player
-      objectReference: {fileID: 0}
-    - target: {fileID: 5430377791670745392, guid: 0616308f626330740800bf213bc4dd38,
-        type: 3}
-      propertyPath: basicAttackRang
-      value: 0.48
-      objectReference: {fileID: 0}
-    - target: {fileID: 5430377791670745392, guid: 0616308f626330740800bf213bc4dd38,
-        type: 3}
-      propertyPath: basicAttackPoint
-      value: 
-      objectReference: {fileID: 2027668770}
-    - target: {fileID: 5430377791670745392, guid: 0616308f626330740800bf213bc4dd38,
-        type: 3}
-      propertyPath: basicAttackRange
-      value: 0.6
-      objectReference: {fileID: 0}
-    - target: {fileID: 5430377791670745392, guid: 0616308f626330740800bf213bc4dd38,
-        type: 3}
-      propertyPath: enemyLayers.m_Bits
-      value: 128
       objectReference: {fileID: 0}
     - target: {fileID: 5430377791670745396, guid: 0616308f626330740800bf213bc4dd38,
         type: 3}
@@ -1488,11 +1444,6 @@ PrefabInstance:
     - target: {fileID: 5430377791670745396, guid: 0616308f626330740800bf213bc4dd38,
         type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5430377791670745398, guid: 0616308f626330740800bf213bc4dd38,
-        type: 3}
-      propertyPath: m_Drag
       value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
