+++ Iso Adventure/Logs/shadercompiler-AssetImportWorker0.log
@@ -1,0 +1,3 @@
+Base path: 'C:/Program Files/2021.1.22f1/Editor/Data', plugins path 'C:/Program Files/2021.1.22f1/Editor/Data/PlaybackEngines'
+Cmd: initializeCompiler
+
