--- Iso Adventure/Logs/shadercompiler-UnityShaderCompiler.exe8.log
@@ -1,3 +1,0 @@
-Base path: 'C:/Program Files/Unity/Hub/Editor/2021.1.22f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2021.1.22f1/Editor/Data/PlaybackEngines'
-Cmd: initializeCompiler
-
